Fixes warnings in verbose mode.
Fixes "Instance variable not defined", "shadowing outer local variable" and
"assigned but unused variable" warnings when using simple_router with rspec and
petty mode turned on.

The reason for wanting these warnings gone in verbose mode is that they get
mixed in with the warnings from the application under test and make it hard to
see what warnings are from the application and what comes from included
libraries.

This isn't an essential patch, it doesn't fix an error. But it would be nice to
have :)